split `utils.py` into `env.py, strings.py, utils.py`
