+++ langchain/utils/utils.py
@@ -2,36 +2,11 @@
 import contextlib
 import datetime
 import importlib
-import os
 from importlib.metadata import version
-from typing import Any, Callable, Dict, List, Optional, Set, Tuple
+from typing import Any, Callable, Optional, Set, Tuple
 
 from packaging.version import parse
 from requests import HTTPError, Response
-
-
-def get_from_dict_or_env(
-    data: Dict[str, Any], key: str, env_key: str, default: Optional[str] = None
-) -> str:
-    """Get a value from a dictionary or an environment variable."""
-    if key in data and data[key]:
-        return data[key]
-    else:
-        return get_from_env(key, env_key, default=default)
-
-
-def get_from_env(key: str, env_key: str, default: Optional[str] = None) -> str:
-    """Get a value from a dictionary or an environment variable."""
-    if env_key in os.environ and os.environ[env_key]:
-        return os.environ[env_key]
-    elif default is not None:
-        return default
-    else:
-        raise ValueError(
-            f"Did not find {key}, please add an environment variable"
-            f" `{env_key}` which contains it, or pass"
-            f"  `{key}` as a named parameter."
-        )
 
 
 def xor_args(*arg_groups: Tuple[str, ...]) -> Callable:
@@ -65,44 +40,6 @@ def raise_for_status_with_text(response: Response) -> None:
         response.raise_for_status()
     except HTTPError as e:
         raise ValueError(response.text) from e
-
-
-def stringify_value(val: Any) -> str:
-    """Stringify a value.
-
-    Args:
-        val: The value to stringify.
-
-    Returns:
-        str: The stringified value.
-    """
-    if isinstance(val, str):
-        return val
-    elif isinstance(val, dict):
-        return "\n" + stringify_dict(val)
-    elif isinstance(val, list):
-        return "\n".join(stringify_value(v) for v in val)
-    else:
-        return str(val)
-
-
-def stringify_dict(data: dict) -> str:
-    """Stringify a dictionary.
-
-    Args:
-        data: The dictionary to stringify.
-
-    Returns:
-        str: The stringified dictionary.
-    """
-    text = ""
-    for key, value in data.items():
-        text += key + ": " + stringify_value(value) + "\n"
-    return text
-
-
-def comma_list(items: List[Any]) -> str:
-    return ", ".join(str(item) for item in items)
 
 
 @contextlib.contextmanager
